+++ app/assets/javascripts/spree/backend/solidus_avatax_certified.js
@@ -1,5 +1,6 @@
 //= require spree/backend/use_code_picker
 //= require spree/address_validator
 
+Spree.routes = Spree.routes || {}
 Spree.routes.use_code_search = Spree.pathFor("admin/avalara_entity_use_codes")
 Spree.routes.validate_address = Spree.pathFor("admin/avatax_settings/validate_address")