+++ db/migrate/20160624154423_create_spree_avalara_entity_use_codes.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateSpreeAvalaraEntityUseCodes < SolidusSupport::Migration[4.2]
+class CreateSpreeAvalaraEntityUseCodes < ActiveRecord::Migration[4.2]
   def change
     create_table :spree_avalara_entity_use_codes do |t|
       t.string :use_code